added permissions so certain actions like delete are only accessible to users with permissions

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require("jsonwebtoken");
 const { randomBytes } = require("crypto");
 const { promisify } = require("util");
 const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -50,8 +51,16 @@ const Mutations = {
 
     //1. find the item
 
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user {id}}`);
     //2. check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You Do not have Permission");
+    }
     //3. Delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -179,6 +188,39 @@ const Mutations = {
 
     // 8. return new user
     return updatedUser;
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. check to see if user is logged in
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in");
+    }
+
+    // 2. query the current user
+    const currentUser = await ctx.db.query.user({
+      where: {
+        id: ctx.request.userId
+      },
+      info
+    });
+    currentUser.permissions = ctx.request.user.permissions;
+    // 3. check to see if user has permission
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+
+    // 4. update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
